pyproject.toml
[tool.poetry]
name = "discord-integrator"
version = "0.1.0"
description = ""
authors = ["Maruo.S <raspi-maru2004@outlook.jp>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
py-cord = "^2.5.0"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
watchfiles = "^0.21.0"
tortoise-cli = "^0.1.2"


[tool.poetry.group.postgres.dependencies]
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}


[tool.aerich]
tortoise_orm = "main.DB_CONFIG"
location = "./migrations"
src_folder = "./src"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
